requirements.txt
streamlit
uvicorn
fastapi
langgraph
python-dotenv
google-generativeai
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
